analyzer_v4/requirements_V4.txt
akshare>=1.10.0
pandas>=1.3.0
openpyxl>=3.0.0
numpy>=1.21.0
requests>=2.26.0
python-dateutil>=2.8.2
pytz>=2021.1 
yfinance>=0.2.36 